For GPRC logger: moving Log messages classes to dedicated file + changing the color of some log messages.
